pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "boto3>=1.38.27",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "plotly>=6.1.2",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.41",
    "streamlit-authenticator>=0.4.2",
    "streamlit>=1.45.1",
    "botocore>=1.38.27",
]
